Skip image building if matching image already exists
This change allows a build step to pass on an existing image if it
matches, avoiding the need to rebuild, tag, and push the image up to
Docker Hub if it's actually identical. For some images such as our
dependency sub-layer images, this can cut the time the build step runs
in by 2/3rds for an unchanged build. It has the side benefit that
existing build agent machines that already had the prior cached image
can continue to reuse it, further speeding up builds.